+++ monkey/monkey_island/cc/ui/src/components/pages/RunMonkeyPage/RunManually/LocalManualRunOptions.js
@@ -20,9 +20,7 @@ const getContents = (props) => {
 
   const osTypes = {
     [OS_TYPES.WINDOWS_64]: 'Windows 64bit',
-    [OS_TYPES.WINDOWS_32]: 'Windows 32bit',
-    [OS_TYPES.LINUX_64]: 'Linux 64bit',
-    [OS_TYPES.LINUX_32]: 'Linux 32bit'
+    [OS_TYPES.LINUX_64]: 'Linux 64bit'
   }
 
   const [osType, setOsType] = useState(OS_TYPES.WINDOWS_64);
@@ -48,11 +46,11 @@ const getContents = (props) => {
   }
 
   function generateCommands() {
-    if (osType === OS_TYPES.WINDOWS_64 || osType === OS_TYPES.WINDOWS_32) {
-      return [{type: 'Powershell', command: GenerateLocalWindowsPowershell(selectedIp, osType, customUsername)}]
+    if (osType === OS_TYPES.WINDOWS_64) {
+      return [{type: 'Powershell', command: GenerateLocalWindowsPowershell(selectedIp, customUsername)}]
     } else {
-      return [{type: 'CURL', command: GenerateLocalLinuxCurl(selectedIp, osType, customUsername)},
-        {type: 'WGET', command: GenerateLocalLinuxWget(selectedIp, osType, customUsername)}]
+      return [{type: 'CURL', command: GenerateLocalLinuxCurl(selectedIp, customUsername)},
+        {type: 'WGET', command: GenerateLocalLinuxWget(selectedIp, customUsername)}]
     }
   }